requirements.txt
asgiref==3.7.2
aws-psycopg2==1.3.8 #on mac M1/M2: mamba install psycopg2 --> DON'T USE PIP
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.7
Django<4.2
django-crispy-forms==2.0
django-formtools==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
